+++ SolastaUnfinishedBusiness/Displays/CreditsDisplay.cs
@@ -63,7 +63,7 @@ internal static class CreditsDisplay
 
         if (IsUnityExplorerInstalled && !IsUnityExplorerEnabled)
         {
-            UI.ActionButton("Unity Explorer UI".Bold().Khaki(), EnableUnityExplorerUi, UI.Width(150));
+            UI.ActionButton("Unity Explorer UI".Bold().Khaki(), EnableUnityExplorerUi, UI.Width((float)150));
             UI.Label();
         }
 
@@ -81,8 +81,8 @@ internal static class CreditsDisplay
             {
                 using (UI.HorizontalScope())
                 {
-                    UI.Label(author.Orange(), UI.Width(150));
-                    UI.Label(content, UI.Width(600));
+                    UI.Label(author.Orange(), UI.Width((float)150));
+                    UI.Label(content, UI.Width((float)600));
                 }
             }
         }
